fix: move icon client transfer to ClientTest.php
