+++ tests/DevLiveServer/FileTransferTest.php
@@ -42,10 +42,6 @@ class FileTransferTest extends TestCase
 
     private int $test_cid;
 
-    private string $user_test_active;
-
-    private string $ts3_unit_test_userName;
-
     public function setUp(): void
     {
         //proof test active
@@ -59,8 +55,6 @@ class FileTransferTest extends TestCase
             $this->password = str_replace('DEV_LIVE_SERVER_QUERY_USER_PASSWORD=', '', preg_replace('#\n(?!\n)#', '', $env[6]));
             $this->serverPort = str_replace('DEV_LIVE_SERVER_UNIT_TEST_SERVER_PORT=', '', preg_replace('#\n(?!\n)#', '', $env[12]));
             $this->ts3_unit_test_channel_name = str_replace('DEV_LIVE_SERVER_UNIT_TEST_CHANNEL=', '', preg_replace('#\n(?!\n)#', '', $env[7]));
-            $this->user_test_active = str_replace('DEV_LIVE_SERVER_UNIT_TEST_USER_ACTIVE=', '', preg_replace('#\n(?!\n)#', '', $env[8]));
-            $this->ts3_unit_test_userName = str_replace('DEV_LIVE_SERVER_UNIT_TEST_USER=', '', preg_replace('#\n(?!\n)#', '', $env[9]));
         } else {
             $this->active = 'false';
         }
@@ -395,59 +389,6 @@ class FileTransferTest extends TestCase
         $ts3_VirtualServer->iconDelete($iconId);
         $this->unset_play_test_channel($ts3_VirtualServer);
         $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
-        $this->assertFalse($ts3_VirtualServer->getAdapter()->getTransport()->isConnected());
-    }
-
-    /**
-     * @throws TransportException
-     * @throws ServerQueryException
-     * @throws AdapterException
-     * @throws FileTransferException
-     * @throws HelperException
-     */
-    public function test_can_download_client_icon()
-    {
-        if ($this->user_test_active == 'false' || $this->active == 'false') {
-            $this->markTestSkipped('DevLiveServer ist not active');
-        }
-
-        $ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri);
-
-        $iconFile = getcwd().DIRECTORY_SEPARATOR.'tests'.DIRECTORY_SEPARATOR.'testsources'.DIRECTORY_SEPARATOR.'icons'.DIRECTORY_SEPARATOR.'upload'.DIRECTORY_SEPARATOR.'Voice.png';
-        $this->assertFileExists($iconFile);
-        $this->assertGreaterThan(0, filesize($iconFile));
-
-        $iconId = $ts3_VirtualServer->iconUpload($iconFile);
-        $signedIconId = $iconId > 0x7FFFFFFF ? $iconId - 0x100000000 : $iconId;
-        $iconName = 'icon_'.$iconId;
-
-        $client = $ts3_VirtualServer->clientGetByName($this->ts3_unit_test_userName);
-        $client->permAssign(['i_icon_id'], $signedIconId);
-
-        $permissions = $client->permList(true);
-
-        $this->assertArrayHasKey('i_icon_id', $permissions);
-        $this->assertSame($signedIconId, (int) $permissions['i_icon_id']['permvalue']);
-
-        $ts3_VirtualServer->clientListReset();
-        $client = $ts3_VirtualServer->clientGetByName($this->ts3_unit_test_userName);
-        $content = $client->iconDownload();
-
-        $this->assertNotNull($content);
-        $this->assertGreaterThan(0, strlen($content->toString()));
-
-        $downloadPath = getcwd().$this->testPath.DIRECTORY_SEPARATOR.'icons'.DIRECTORY_SEPARATOR.'download';
-        $targetFile = $downloadPath.DIRECTORY_SEPARATOR.$iconName.'.png';
-
-        file_put_contents($targetFile, $content->toString());
-
-        $this->assertFileExists($targetFile);
-        $this->assertGreaterThan(0, filesize($targetFile));
-
-        $client->permRemove(['i_icon_id']);
-        $ts3_VirtualServer->iconDelete($iconId);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
-
         $this->assertFalse($ts3_VirtualServer->getAdapter()->getTransport()->isConnected());
     }
